requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
idna==3.11
    # via requests
numpy==2.2.6
    # via pandas
pandas==2.3.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
requests==2.32.5
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
